sherlock-tensorflow/requirements.txt
absl-py==0.7.0
amqp==2.4.1
astor==0.7.1
billiard==3.5.0.5
boto3==1.9.88
botocore==1.12.88
celery==4.2.1
docutils==0.14
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kombu==4.3.0
Markdown==3.0.1
numpy==1.16.1
Pillow==5.4.1
protobuf==3.6.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
redis==3.1.0
s3transfer==0.1.13
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
